pyproject.toml
[project]
name = "mini-siem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.119.0",
    "orjson>=3.11.3",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "coverage>=7.11.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.0",
]
